Disk Cache: design an API to disable GC collection

+++ src/lib/disk_cache/filesystem/disk_cache.ml
@@ -10,36 +10,44 @@ module Make (B : sig
   include Binable.S
 end) =
 struct
-  type t = string * int ref
+  type t = { root : string; next_idx : int ref; eviction_freezed : bool ref }
+
+  type persistence = int [@@deriving bin_io_unversioned]
 
   type id = { idx : int }
 
-  let initialize path ~logger =
+  let initialize path ~logger ?persistence:(next_idx : persistence = 0) () =
     Async.Deferred.Result.map (Disk_cache_utils.initialize_dir path ~logger)
-      ~f:(fun path -> (path, ref 0))
+      ~f:(fun root ->
+        { root; next_idx = ref next_idx; eviction_freezed = ref false } )
 
   let path root i = root ^ Filename.dir_sep ^ Int.to_string i
 
-  let get ((root, _) : t) (id : id) : B.t =
+  let get ({ root; _ } : t) (id : id) : B.t =
     (* Read from the file. *)
     In_channel.with_file ~binary:true (path root id.idx) ~f:(fun chan ->
         let str = In_channel.input_all chan in
         Binable.of_string (module B) str )
 
-  let put ((root, next_idx) : t) x : id =
+  let freeze_eviction { eviction_freezed; next_idx; _ } =
+    eviction_freezed := true ;
+    !next_idx
+
+  let put ({ root; next_idx; eviction_freezed } : t) x : id =
     let idx = !next_idx in
     incr next_idx ;
     let res = { idx } in
     (* When this reference is GC'd, delete the file. *)
     Core.Gc.Expert.add_finalizer_last_exn res (fun () ->
-        (* Ignore errors: if a directory is deleted, it's ok. *)
-        try Core.Unix.unlink (path root idx) with _ -> () ) ;
+        if not !eviction_freezed then
+          (* Ignore errors: if a directory is deleted, it's ok. *)
+          try Core.Unix.unlink (path root idx) with _ -> () ) ;
     (* Write the proof to the file. *)
     Out_channel.with_file ~binary:true (path root idx) ~f:(fun chan ->
         Out_channel.output_string chan @@ Binable.to_string (module B) x ) ;
     res
 
-  let count ((path, _) : t) = Sys.ls_dir path |> List.length
+  let count ({ root; _ } : t) = Sys.ls_dir root |> List.length
 end
 
 let%test_module "disk_cache filesystem" =
